% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursory.R
\name{cursory_all}
\alias{cursory_all}
\title{Cursory Functions}
\usage{
cursory_all(.tbl, .funs, ..., var.name = "Variable")
}
\arguments{
\item{.tbl}{A \code{tbl} object.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}

\item{...}{Additional arguments for the function calls in
\code{.funs}. These are evaluated only once, with \link[rlang:tidy-dots]{tidy dots} support.}
}
\description{
look Cursory functions act like the dplyr `summarise_(all|at|if)`
functions with an important difference, they put the variable name
in a column and for each function passed in it puts the value in
it's own column.
}
\details{
* `cursory_all()` is the analog of [dplyr::summarise_all()]
* `cursory_at()` is the analog of [dplyr::summarise_at()]
* `cursory_if()` is the analog of [dplyr::summarise_if()]
}
