% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontrepeat.R
\name{dontrepeat}
\alias{dontrepeat}
\title{Indicate that when printing repeat values should be hidden.}
\usage{
dontrepeat(x, replace.with = "")
}
\arguments{
\item{x}{a vector}

\item{replace.with}{what to replace the value with.}
}
\description{
Indicate that when printing repeat values should be hidden.
}
\examples{
library(dplyr)
library(tibble)
x <- cursory_all(group_by(iris, Species), lst(mean, sd))
x <- as.data.frame(arrange(x, Species))
print(x)
x[[1]] <- dontrepeat(x[[1]], replace.with='"')
print(x)
}
