% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursory.R
\name{cursory_all}
\alias{cursory_all}
\alias{cursory_at}
\alias{cursory_if}
\title{Cursory Functions}
\usage{
cursory_all(.tbl, .funs, ..., var.name = "Variable")

cursory_at(.tbl, .vars, .funs, ..., var.name = "Variable")

cursory_if(.tbl, .predicate, .funs, ..., var.name = "Variable")
}
\arguments{
\item{.tbl}{A \code{tbl} object.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}

\item{...}{Additional arguments for the function calls in
\code{.funs}. These are evaluated only once, with \link[rlang:tidy-dots]{tidy dots} support.}

\item{var.name}{Name of the column with variable names.}

\item{.vars}{A list of columns generated by \code{\link[=vars]{vars()}},
a character vector of column names, a numeric vector of column
positions, or \code{NULL}.}

\item{.predicate}{A predicate function to be applied to the columns
or a logical vector. The variables for which \code{.predicate} is or
returns \code{TRUE} are selected. This argument is passed to
\code{\link[rlang:as_function]{rlang::as_function()}} and thus supports quosure-style lambda
functions and strings representing function names.}
}
\description{
Cursory functions act like the dplyr \code{summarise_(all|at|if)}
functions with an important difference, they put the variable name
in a column and for each function passed in it puts the value in
it's own column.
}
\details{
\itemize{
\item \code{cursory_all()} is the analog of \code{\link[dplyr:summarise_all]{dplyr::summarise_all()}}
\item \code{cursory_at()} is the analog of \code{\link[dplyr:summarise_at]{dplyr::summarise_at()}}
\item \code{cursory_if()} is the analog of \code{\link[dplyr:summarise_if]{dplyr::summarise_if()}}
}
}
